requirements.txt
Django==2.2.5
djangorestframework==3.10.3
Pygments==2.4.2
pytz==2019.2
sqlparse==0.3.0